requirements.txt
imapclient>=2.3.1
msal>=1.26.0
openpyxl>=3.1.0
pyyaml>=6.0
python-dateutil>=2.8.0
flask>=3.0.0
apscheduler>=3.10.0
anthropic>=0.39.0
google-genai>=1.0.0
groq>=0.12.0
